+++ js/label/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/label
 
+## 0.7.1
+
+### Dependency updates
+
+- @gradio/statustracker@0.15.1
+- @gradio/icons@0.16.0
+- @gradio/atoms@0.26.0
+
 ## 0.7.0
 
 ### Features