+++ js/slider/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/slider
 
+## 0.8.1
+
+### Dependency updates
+
+- @gradio/statustracker@0.15.1
+- @gradio/atoms@0.26.0
+
 ## 0.8.0
 
 ### Features