+++ js/vibeeditor/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/vibeeditor
 
+## 0.3.12
+
+### Dependency updates
+
+- @gradio/statustracker@0.15.1
+- @gradio/atoms@0.26.0
+- @gradio/client@2.3.1
+
 ## 0.3.11
 
 ### Dependency updates